Add files via upload
This code employs the SMOTE technique to oversample the dataset in the presence of minority datapoints. SMOTE is favored over traditional upsampling because, in the case of upsampling, similar datapoints are added to the minority datapoints, resulting in minimal alteration to the variance within the data. In contrast, SMOTE enhances oversampling by introducing new data points, thereby increasing the variance in the data.