Add files via upload
This  implements multiple regression analysis on the California Housing dataset. It leverages Scikit-Learn for modeling and calculates essential performance metrics to evaluate the regression model's accuracy. Furthermore, the script demonstrates the use of pickling to serialize and save the trained model, enabling convenient future deployments and predictions. Whether you're interested in regression modeling or need a reference for model persistence, this script offers a practical and self-contained solution.